mf5/requirements_mf5.txt
torch
torchvision
tqdm
PyYAML
numpy
Pillow
kornia
opencv-python
albumentations
wandb
